Split timeline UI into separate file
